// src/gen/models/createFolderRequestDTO.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * STORE ME API
 * Hệ thống lưu trữ file cho phép người dùng upload, download và quản lý file của mình.

 * OpenAPI spec version: 1.0 beta
 */

export interface CreateFolderRequestDTO {
  parentId?: number;
  /**
   * @minLength 0
   * @maxLength 255
   */
  name: string;
  locked?: boolean;
  hidden?: boolean;
}